hermes_reader/requirements.txt
async-timeout==4.0.3
pyaes==1.6.1
pyasn1==0.6.0
python-dotenv==1.0.1
redis==5.0.6
rsa==4.9
Telethon==1.36.0
